requirements.txt
# Core
agent-framework>=0.1.0
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b7

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Azure SDKs
azure-search-documents>=11.6.0b8
azure-storage-blob>=12.20.0
